requirements.txt
num2words
faster-whisper
more-itertools
tiktoken
cython
torch
torchaudio
soundfile
ffmpeg-python
librosa==0.8.0
scikit-learn
numba
inflect
tqdm
anyascii
pyyaml
fsspec
aiohttp
packaging
pysbd
umap-learn
pandas
jiwer
pymongo
scipy
pydub
opencv-python
torchvision
mediapipe
opencv_python